Initial update to new design
Updates the `@redocly/developer-portal` to the latest version and cleans up `package.json`. Removes the `package-lock.json` as this repository is managed with `yarn` rather than `npm`.

Updates the theme file to more closely match the intended design.

Overrides the default `NavBar` and `Footer` components so that we can achieve the intended look and behavior.

Introduces the `IndexHero` component used for the big hero in the index page.